backend/requirements.txt
arabic-reshaper==3.0.0
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
PyJWT==2.6.0
PyMySQL==1.0.2
six==1.16.0
SQLAlchemy==2.0.4
termcolor==2.2.0
typing_extensions==4.5.0
Werkzeug==2.2.3
windows-curses==2.3.1
